+++ app/src/main/java/com/numero/material_gallery/components/sheet/ModalBottomSheetActivity.kt
@@ -1,25 +1,20 @@
-package com.numero.material_gallery.activity
+package com.numero.material_gallery.components.sheet
 
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
-import android.view.View
-import android.widget.LinearLayout
 import androidx.appcompat.app.AppCompatActivity
-import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.numero.material_gallery.R
 import com.numero.material_gallery.fragment.ThemeInfoBottomSheetDialog
 import com.numero.material_gallery.repository.ConfigRepository
-import kotlinx.android.synthetic.main.activity_bottom_sheet.*
+import kotlinx.android.synthetic.main.activity_modal_bottom_sheet.*
 import org.koin.android.ext.android.inject
 
-class BottomSheetActivity : AppCompatActivity(R.layout.activity_bottom_sheet) {
+class ModalBottomSheetActivity : AppCompatActivity(R.layout.activity_modal_bottom_sheet) {
 
     private val configRepository by inject<ConfigRepository>()
-
-    private lateinit var behavior: BottomSheetBehavior<LinearLayout>
 
     override fun onCreate(savedInstanceState: Bundle?) {
         setTheme(configRepository.themeRes)
@@ -31,22 +26,7 @@ class BottomSheetActivity : AppCompatActivity(R.layout.activity_bottom_sheet) {
         }
 
         showBottomSheetButton.setOnClickListener {
-            behavior.state = BottomSheetBehavior.STATE_EXPANDED
-        }
-        draggableSwitch.setOnCheckedChangeListener { _, isChecked ->
-            behavior.isDraggable = isChecked
-        }
-
-        behavior = BottomSheetBehavior.from(bottomSheetLayout).apply {
-            state = BottomSheetBehavior.STATE_HIDDEN
-            addBottomSheetCallback(object : BottomSheetBehavior.BottomSheetCallback() {
-                override fun onSlide(p0: View, p1: Float) {
-                }
-
-                override fun onStateChanged(view: View, state: Int) {
-                    showBottomSheetButton.isEnabled = state == BottomSheetBehavior.STATE_HIDDEN
-                }
-            })
+            BottomSheetModalFragment.newInstance().showIfNeeded(supportFragmentManager)
         }
     }
 
@@ -70,6 +50,6 @@ class BottomSheetActivity : AppCompatActivity(R.layout.activity_bottom_sheet) {
     }
 
     companion object {
-        fun createIntent(context: Context): Intent = Intent(context, BottomSheetActivity::class.java)
+        fun createIntent(context: Context): Intent = Intent(context, ModalBottomSheetActivity::class.java)
     }
 }
