+++ app/src/main/res/layout/activity_modal_bottom_sheet.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.ModalBottomSheetActivity">
+    tools:context=".components.sheet.ModalBottomSheetActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
